BACKEND/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-decouple==3.8
slowapi==0.1.9
jinja2==3.1.2
fastapi-mail==1.4.1
python-dotenv==1.0.0
email-validator==2.1.0
user-agents==2.2.0
apscheduler==3.11.1
psycopg2-binary==2.9.9
sib-api-v3-sdk
pyotp==2.9.0
qrcode==7.4.2
